pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cerebrum"
version = "0.1.0"
description = "Personal Knowledge Refinement System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "rich>=13.7.0",
    "python-frontmatter>=1.0.0",
    "pyyaml>=6.0.1",
    "pypdf>=3.17.0",
]

[project.optional-dependencies]
local = [
    "ollama>=0.1.6",
]
cloud = [
    "google-genai>=1.15.0",
]
full = [
    "ollama>=0.1.6",
    "google-genai>=1.15.0",
    "chromadb>=0.4.18",
    "sentence-transformers>=2.2.2",
    "networkx>=3.2",
]

[project.scripts]
cerebrum = "cerebrum.cli:cli"

[tool.setuptools.packages.find]
include = ["cerebrum*"]
